add abstract procedure
