+++ confignode/src/main/java/org/apache/iotdb/confignode/manager/ProcedureManager.java
@@ -150,6 +150,7 @@ public class ProcedureManager {
         .forEach(
             tDataNodeLocation -> {
               this.executor.submitProcedure(new RemoveDataNodeProcedure(tDataNodeLocation));
+              LOGGER.info("Submit to remove data node procedure, {}", tDataNodeLocation);
             });
     return true;
   }
