+++ confignode/src/main/java/org/apache/iotdb/confignode/procedure/impl/RemoveConfigNodeProcedure.java
@@ -22,10 +22,8 @@ package org.apache.iotdb.confignode.procedure.impl;
 import org.apache.iotdb.common.rpc.thrift.TConfigNodeLocation;
 import org.apache.iotdb.commons.exception.runtime.ThriftSerDeException;
 import org.apache.iotdb.commons.utils.ThriftConfigNodeSerDeUtils;
-import org.apache.iotdb.confignode.procedure.StateMachineProcedure;
 import org.apache.iotdb.confignode.procedure.env.ConfigNodeProcedureEnv;
 import org.apache.iotdb.confignode.procedure.exception.ProcedureException;
-import org.apache.iotdb.confignode.procedure.state.ProcedureLockState;
 import org.apache.iotdb.confignode.procedure.state.RemoveConfigNodeState;
 import org.apache.iotdb.confignode.procedure.store.ProcedureFactory;
 
@@ -37,8 +35,7 @@ import java.io.IOException;
 import java.nio.ByteBuffer;
 
 /** remove config node procedure */
-public class RemoveConfigNodeProcedure
-    extends StateMachineProcedure<ConfigNodeProcedureEnv, RemoveConfigNodeState> {
+public class RemoveConfigNodeProcedure extends AbstractNodeProcedure<RemoveConfigNodeState> {
   private static final Logger LOG = LoggerFactory.getLogger(RemoveConfigNodeProcedure.class);
   private static final int retryThreshold = 5;
 
@@ -102,37 +99,6 @@ public class RemoveConfigNodeProcedure
   @Override
   protected boolean isRollbackSupported(RemoveConfigNodeState state) {
     return true;
-  }
-
-  @Override
-  protected ProcedureLockState acquireLock(ConfigNodeProcedureEnv configNodeProcedureEnv) {
-    configNodeProcedureEnv.getSchedulerLock().lock();
-    try {
-      if (configNodeProcedureEnv.getNodeLock().tryLock(this)) {
-        LOG.info("{} acquire lock.", getProcId());
-        return ProcedureLockState.LOCK_ACQUIRED;
-      }
-      configNodeProcedureEnv.getNodeLock().waitProcedure(this);
-      LOG.info("{} wait for lock.", getProcId());
-      return ProcedureLockState.LOCK_EVENT_WAIT;
-    } finally {
-      configNodeProcedureEnv.getSchedulerLock().unlock();
-    }
-  }
-
-  @Override
-  protected void releaseLock(ConfigNodeProcedureEnv configNodeProcedureEnv) {
-    configNodeProcedureEnv.getSchedulerLock().lock();
-    try {
-      LOG.info("{} release lock.", getProcId());
-      if (configNodeProcedureEnv.getNodeLock().releaseLock(this)) {
-        configNodeProcedureEnv
-            .getNodeLock()
-            .wakeWaitingProcedures(configNodeProcedureEnv.getScheduler());
-      }
-    } finally {
-      configNodeProcedureEnv.getSchedulerLock().unlock();
-    }
   }
 
   @Override
